requirements.txt
fastapi
uvicorn
python-dotenv 
google-cloud-documentai 
python-multipart
